Move all authentication information to httponly cookies
Instead of sending the JWT token in the Auth header (which does not seem to have the intended behavior in some browsers), send
it in an HttpOnly cookie like the refresh token. Similarly, set a non-httponly cookie with the user information for the SPA
to access for request building to other services. This will require other services to verify the unsecure user information
against the information included in the JWT.